+++ zcash_client_sqlite/src/testing.rs
@@ -54,7 +54,7 @@ use zcash_primitives::{
     consensus::{self, BlockHeight, Network, NetworkUpgrade, Parameters},
     memo::{Memo, MemoBytes},
     transaction::{
-        components::amount::NonNegativeAmount,
+        components::{amount::NonNegativeAmount, sapling::zip212_enforcement},
         fees::{zip317::FeeError as Zip317FeeError, FeeRule, StandardFeeRule},
         Transaction, TxId,
     },
@@ -784,11 +784,8 @@ pub(crate) fn fake_compact_block<P: consensus::Parameters>(
 
     // Create a fake Note for the account
     let mut rng = OsRng;
-    let rseed = generate_random_rseed(
-        consensus::sapling_zip212_enforcement(params, height),
-        &mut rng,
-    );
-    let note = Note::from_parts(to, NoteValue::from(value), rseed);
+    let rseed = generate_random_rseed(zip212_enforcement(params, height), &mut rng);
+    let note = Note::from_parts(to, NoteValue::from_raw(value.into_u64()), rseed);
     let encryptor = sapling_note_encryption(
         Some(dfvk.fvk().ovk),
         note.clone(),
@@ -883,7 +880,7 @@ pub(crate) fn fake_compact_block_spending<P: consensus::Parameters>(
     value: NonNegativeAmount,
     initial_sapling_tree_size: u32,
 ) -> CompactBlock {
-    let zip212_enforcement = consensus::sapling_zip212_enforcement(params, height);
+    let zip212_enforcement = zip212_enforcement(params, height);
     let mut rng = OsRng;
     let rseed = generate_random_rseed(zip212_enforcement, &mut rng);
 
@@ -897,7 +894,11 @@ pub(crate) fn fake_compact_block_spending<P: consensus::Parameters>(
 
     // Create a fake Note for the payment
     ctx.outputs.push({
-        let note = Note::from_parts(to, NoteValue::from(value), rseed);
+        let note = Note::from_parts(
+            to,
+            sapling::value::NoteValue::from_raw(value.into_u64()),
+            rseed,
+        );
         let encryptor = sapling_note_encryption(
             Some(dfvk.fvk().ovk),
             note.clone(),
@@ -921,7 +922,7 @@ pub(crate) fn fake_compact_block_spending<P: consensus::Parameters>(
         let rseed = generate_random_rseed(zip212_enforcement, &mut rng);
         let note = Note::from_parts(
             change_addr,
-            NoteValue::from((in_value - value).unwrap()),
+            NoteValue::from_raw((in_value - value).unwrap().into_u64()),
             rseed,
         );
         let encryptor = sapling_note_encryption(