+++ zcash_primitives/src/lib.rs
@@ -18,16 +18,14 @@
 #![allow(clippy::single_component_path_imports)]
 
 pub mod block;
-pub mod consensus;
-pub mod constants;
+pub use zcash_protocol::consensus;
+pub use zcash_protocol::constants;
 pub mod legacy;
-pub mod memo;
+pub use zcash_protocol::memo;
 pub mod merkle_tree;
 use sapling;
 pub mod transaction;
 pub use zip32;
 #[cfg(feature = "zfuture")]
 pub mod extensions;
-#[cfg(feature = "local-consensus")]
-pub mod local_consensus;
 pub mod zip339;